Moved parser_state into its own header
